Migrate 'internalization' demo to NIM.
