+++ internationalization/index.html
@@ -201,9 +201,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 			</div>
 		</div>
 
-		<script src="./dist/translations/de.js"></script>
-		<script src="./dist/translations/ar.js"></script>
-		<script src="./dist/translations/ja.js"></script>
-		<script src="./dist/main.js"></script>
+		<script type="module" src="index.ts"></script>
 	</body>
 </html>
