+++ internationalization/package.json
@@ -1,60 +1,21 @@
 {
 	"name": "internationalization",
-	"version": "0.0.1",
-	"main": "index.js",
+	"version": "0.1.0",
 	"repository": "https://github.com/ckeditor/ckeditor5-demos",
 	"author": "CKSource (http://cksource.com/)",
 	"license": "GPL-2.0-or-later",
 	"private": true,
-	"devDependencies": {
-		"@babel/core": "^7.20.7",
-		"@ckeditor/ckeditor5-adapter-ckfinder": "40.2.0",
-		"@ckeditor/ckeditor5-alignment": "40.2.0",
-		"@ckeditor/ckeditor5-autoformat": "40.2.0",
-		"@ckeditor/ckeditor5-basic-styles": "40.2.0",
-		"@ckeditor/ckeditor5-block-quote": "40.2.0",
-		"@ckeditor/ckeditor5-ckfinder": "40.2.0",
-		"@ckeditor/ckeditor5-cloud-services": "40.2.0",
-		"@ckeditor/ckeditor5-core": "40.2.0",
-		"@ckeditor/ckeditor5-dev-translations": "^39.5.0",
-		"@ckeditor/ckeditor5-dev-utils": "^39.5.0",
-		"@ckeditor/ckeditor5-editor-classic": "40.2.0",
-		"@ckeditor/ckeditor5-engine": "40.2.0",
-		"@ckeditor/ckeditor5-enter": "40.2.0",
-		"@ckeditor/ckeditor5-essentials": "40.2.0",
-		"@ckeditor/ckeditor5-find-and-replace": "40.2.0",
-		"@ckeditor/ckeditor5-heading": "40.2.0",
-		"@ckeditor/ckeditor5-highlight": "40.2.0",
-		"@ckeditor/ckeditor5-horizontal-line": "40.2.0",
-		"@ckeditor/ckeditor5-image": "40.2.0",
-		"@ckeditor/ckeditor5-indent": "40.2.0",
-		"@ckeditor/ckeditor5-language": "40.2.0",
-		"@ckeditor/ckeditor5-link": "40.2.0",
-		"@ckeditor/ckeditor5-list": "40.2.0",
-		"@ckeditor/ckeditor5-media-embed": "40.2.0",
-		"@ckeditor/ckeditor5-paragraph": "40.2.0",
-		"@ckeditor/ckeditor5-remove-format": "40.2.0",
-		"@ckeditor/ckeditor5-special-characters": "40.2.0",
-		"@ckeditor/ckeditor5-table": "40.2.0",
-		"@ckeditor/ckeditor5-theme-lark": "40.2.0",
-		"@ckeditor/ckeditor5-typing": "40.2.0",
-		"@ckeditor/ckeditor5-ui": "40.2.0",
-		"@ckeditor/ckeditor5-undo": "40.2.0",
-		"@ckeditor/ckeditor5-upload": "40.2.0",
-		"@ckeditor/ckeditor5-utils": "40.2.0",
-		"css-loader": "^5.2.7",
-		"postcss": "^8.4.19",
-		"postcss-loader": "^4.3.0",
-		"raw-loader": "^4.0.2",
-		"style-loader": "^2.0.0",
-		"terser-webpack-plugin": "^4.2.3",
-		"webpack": "^5.75.0",
-		"webpack-cli": "^4.10.0",
-		"webpack-dev-server": "^4.11.1"
+	"type": "module",
+	"engines": {
+		"node": ">=18.0.0"
 	},
 	"scripts": {
-		"build": "webpack --mode production",
-		"build-dev": "webpack --mode development",
-		"start": "webpack serve --open /index.html --mode development"
+		"dev": "vite --open",
+		"build": "vite build",
+		"preview": "vite preview"
+	},
+	"devDependencies": {
+		"ckeditor5": "nightly",
+		"vite": "^5.0.0"
 	}
 }
